+++ src/checker/get_next_line_utils.c
@@ -6,7 +6,7 @@
 /*   By: anaciri <anaciri@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/12/31 13:30:25 by anaciri           #+#    #+#             */
-/*   Updated: 2022/06/06 00:32:47 by anaciri          ###   ########.fr       */
+/*   Updated: 2022/06/06 02:24:25 by anaciri          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -66,10 +66,10 @@ char	*ft_strjoin(char *s1, char *s2)
 	return (k);
 }
 
-char	*ft_substr(char *s, unsigned int start, size_t len)
+char	*ft_substr(char *s, int start, int len)
 {
 	char	*p;
-	size_t	i;
+	int	i;
 
 	if (s == NULL)
 		return (NULL);
